added ISI.Extensions.AWS

+++ src/ISI.Extensions.GoDaddy/SerializableModels/DomainsApi/DnsRecord.cs
@@ -83,6 +83,6 @@ namespace ISI.Extensions.GoDaddy.SerializableModels.DomainsApi
 		public string RecordType { get; set; }
 
 		[DataMember(Name = "weight", EmitDefaultValue = false)]
-		public int? Weight { get; set; }
+		public long? Weight { get; set; }
 	}
 }
